move iss tracker code to separate component
