+++ sessions/react-data-fetching/iss-tracker/components/ISSTracker/index.js
@@ -1,0 +1,46 @@
+import { useEffect, useState } from "react";
+import Controls from "../Controls/index";
+import Map from "../Map/index";
+
+const URL = "https://api.wheretheiss.at/v1/satellites/25544";
+
+export default function ISSTracker() {
+  const [coords, setCoords] = useState({
+    longitude: 0,
+    latitude: 0,
+  });
+
+  async function getISSCoords() {
+    try {
+      const response = await fetch(URL);
+      if (response.ok) {
+        const data = await response.json();
+        setCoords({ longitude: data.longitude, latitude: data.latitude });
+        console.log(data);
+      }
+    } catch (error) {
+      console.error(error);
+    }
+  }
+
+  useEffect(() => {
+    const timer = setInterval(() => {
+      getISSCoords();
+    }, 5000);
+
+    return () => {
+      clearInterval(timer);
+    };
+  }, []);
+
+  return (
+    <main>
+      <Map longitude={coords.longitude} latitude={coords.latitude} />
+      <Controls
+        longitude={coords.longitude}
+        latitude={coords.latitude}
+        onRefresh={getISSCoords}
+      />
+    </main>
+  );
+}
